+++ CHANGELOG.md
@@ -10,6 +10,7 @@ CHANGELOG
 - **New:** Seed for profile post (#16)
 - **New:** Seed for profile post comment (#17)
 - **New:** Seed for profile post reaction (#18)
+- **New:** Seed for profile post comment reaction (#19)
 
 ## 1.1.0 Alpha 1 (`1010011`)